requirements.txt
# Auto-generated from pyproject.toml — keep in sync
# Run: pip install -e ".[dev]" for development instead
fastapi>=0.111
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
alembic>=1.13
playwright>=1.44
pydantic>=2.7
pydantic-settings>=2.3
pyyaml>=6.0
websockets>=12.0

# SaaS
bcrypt>=4.1
PyJWT>=2.8
stripe>=7.0
httpx>=0.26
python-multipart>=0.0.6
